+++ Web_Development_Fundamentals/block_10/day_1/bonus/lookup/lookup.js
@@ -1,0 +1,43 @@
+// Data
+const users = [
+    {
+      login: 'knuth',
+      firstName: 'Donald',
+      lastName: 'Knuth',
+      likes: ['C', 'Unix'],
+    },
+    {
+      login: 'norvig',
+      firstName: 'Peter',
+      lastName: 'Norvig',
+      likes: ['AI', 'Search', 'NASA', 'Mars'],
+    },
+    {
+      login: 'mfowler',
+      firstName: 'Martin',
+      lastName: 'Fowler',
+      likes: ['Design Patterns', 'Refactoring'],
+    },
+    {
+      login: 'kent',
+      firstName: 'Kent',
+      lastName: 'Beck',
+      likes: ['TDD', 'wikis', 'Design Patterns'],
+    },
+];
+  
+// lookup()
+const lookup = (login, property) => {
+// START -- THIS IS WHERE YOUR CODE GOES
+    const user = users.find(user => user.login === login);
+    if (!user) {
+        throw new Error('Could not find user')
+    } else if (!user[property]) {
+        throw new Error('Could not find property')
+    } else {
+        return user[property]
+    }
+// END
+};
+
+module.exports = lookup;
